feat(refactor): pointing to specific definition file
