+++ nri-config-generator/integration-tests/generator_test.go
