+++ nri-config-generator/integration-tests/testdata/config.yml
