+++ nri-config-generator/integration-tests/testdata/config_with_port.yml
