+++ nri-config-generator/integration-tests/testdata/integration_template/nri-powerdns.json.tmpl
