+++ nri-config-generator/integration-tests/testdata/nri-powerdns-definitions.yml
