+++ nri-config-generator/internal/config/config_test.go
