+++ nri-config-generator/internal/generator/config.go
