+++ nri-config-generator/internal/generator/helper.go
