+++ nri-config-generator/internal/generator/helper_test.go
