+++ nri-config-generator/internal/generator/template.go
