+++ nri-config-generator/internal/generator/template_test.go
