+++ nri-config-generator/internal/synthesis/testdata/sample.yml
