+++ nri-config-generator/internal/synthesis/testdata/sample2.yml
