+++ nri-config-generator/internal/tools/tools.go
